server/requirements.txt
flask[async]~=3.0.0
flask_cors~=4.0.0
python-dotenv~=1.0.0
openai~=1.3.7
requests~=2.31.0
azure-search-documents~=11.4.0
pypdf2~=3.0.1
langchain~=0.0.344
pandas~=2.1.3
beautifulsoup4~=4.12.2
azure-common==1.1.28
azure-core==1.29.5
